Changes in the relationship of the entities and added the usecase and the necessary calls in the repositories. The components of the splash screen are also added

+++ app/src/main/java/es/jco/demo/data/server/dto/UserDTO.kt
@@ -10,7 +10,7 @@ data class UserDTOResult(
 
 @Parcelize
 data class UserDTO (
-    @SerializedName("id") var id: Int,
+    @SerializedName("id") var id: Long,
     @SerializedName("name") var name: String?,
     @SerializedName("username") var username: String?,
     @SerializedName("email") var email: String?,
